Added tests for word set list fragment

+++ app/src/main/java/talkapp/org/talkappmobile/component/database/WordRepetitionProgressService.java
@@ -21,7 +21,7 @@ public interface WordRepetitionProgressService {
 
     Sentence getCurrentSentence(int wordSetId);
 
-    List<WordSet> findFinishedWordSetsSortByUpdatedDate(int limit, int olderThenInHours);
+    List<WordSet> findFinishedWordSetsSortByUpdatedDate(long limit, int olderThenInHours);
 
     List<WordSet> findFinishedWordSetsSortByUpdatedDate(int olderThenInHours);
